sidebar code refactoring
