+++ frontend/src/components/layouts/sidebar.tsx
@@ -1,18 +1,15 @@
 import styles from './sidebar.module.scss'
-import React from 'react'
-import { Link, useLocation, useParams, useSearchParams } from 'react-router-dom'
-import Constants from '../../constants'
-import { DBConnection, DBDataModel, DBQuery, User } from '../../data/models'
-import { useAppDispatch, useAppSelector } from '../../redux/hooks'
-import { selectDBConnection, selectDBDataModels, selectDBDQueries } from '../../redux/dbConnectionSlice'
-import { DBConnType, TabType } from '../../data/defaults'
+import { useLocation } from 'react-router-dom'
+import { DBConnection } from '../../data/models'
+import { useAppSelector } from '../../redux/hooks'
+import { selectDBConnection } from '../../redux/dbConnectionSlice'
 import HomeSidebar from './sidebars/homesidebar'
-import { createTab } from '../../redux/tabsSlice'
-import { selectCurrentUser } from '../../redux/currentUserSlice'
+import DatabaseSidebar from './sidebars/dbsidebar'
+import SettingSidebar from './sidebars/settingsidebar'
 
 enum SidebarViewType {
     HOME = "HOME", // home sidebar
-    DATABASE = "DATABASE", // Used to show elements of single database
+    DATABASE = "DATABASE", // Used to show elements of database screen
     SETTINGS = "SETTINGS" // Used to show elements of settings screen
 }
 
@@ -25,28 +22,6 @@ const Sidebar = () => {
             SidebarViewType.DATABASE : (location.pathname.startsWith("/settings")) ? SidebarViewType.SETTINGS : SidebarViewType.HOME
 
     const dbConnection: DBConnection | undefined = useAppSelector(selectDBConnection)
-    const currentUser: User = useAppSelector(selectCurrentUser)
-    const dbDataModels: DBDataModel[] = useAppSelector(selectDBDataModels)
-    const dbQueries: DBQuery[] = useAppSelector(selectDBDQueries)
-
-    const dispatch = useAppDispatch()
-
-
-    const openDataTab = (schema: string, name: string) => {
-        dispatch(createTab({ dbConnId: dbConnection!.id, tabType: TabType.DATA, metadata: { schema, name } }))
-    }
-
-    const openQueryTab = (queryId: string) => {
-        dispatch(createTab({ dbConnId: dbConnection!.id, tabType: TabType.QUERY, metadata: { queryId } }))
-    }
-
-    const openConsoleTab = () => {
-        dispatch(createTab({ dbConnId: dbConnection!.id, tabType: TabType.CONSOLE, metadata: {} }))
-    }
-
-    const openGenerateSQLTab = () => {
-        dispatch(createTab({ dbConnId: dbConnection!.id, tabType: TabType.GENSQL, metadata: {} }))
-    }
 
     return (
         <aside className={"menu " + styles.sidebar}>
@@ -55,140 +30,10 @@ const Sidebar = () => {
                     <HomeSidebar />
                 }
                 {sidebarView === SidebarViewType.DATABASE && dbConnection &&
-                    <React.Fragment>
-                        <p className="menu-label">
-                            Data Models
-                        </p>
-                        <ul className={"menu-list " + styles.menuList}>
-                            {dbDataModels.map((dataModel: DBDataModel) => {
-                                const label = dbConnection.type === DBConnType.POSTGRES ? `${dataModel.schemaName}.${dataModel.name}` : `${dataModel.name}`
-                                return (
-                                    <li key={dataModel.schemaName + dataModel.name}>
-                                        <a onClick={() => openDataTab(dataModel.schemaName ?? "", dataModel.name)}>
-                                            {label}
-                                        </a>
-                                    </li>
-                                )
-                            })}
-                        </ul>
-                        <p className="menu-label">
-                            Queries
-                        </p>
-                        <ul className={"menu-list " + styles.menuList}>
-                            {dbQueries.map((dbQuery: DBQuery) => {
-                                return (
-                                    <li key={dbQuery.id}>
-                                        <a onClick={() => openQueryTab(dbQuery.id)}>
-                                            {dbQuery.name}
-                                        </a>
-                                    </li>
-                                )
-                            })}
-                            <li>
-                                <a onClick={() => openQueryTab("new")}>
-                                    <span className="icon">
-                                        <i className="fas fa-plus-circle"></i>
-                                    </span>
-                                    &nbsp;New Query
-                                </a>
-                            </li>
-                        </ul>
-                        <p className="menu-label">
-                            Toolbox
-                        </p>
-                        <ul className={"menu-list " + styles.menuList}>
-                            <li>
-                                <a onClick={() => openConsoleTab()}>
-                                    <span className="icon">
-                                        <i className="fas fa-terminal"></i>
-                                    </span>
-                                    &nbsp;Console
-                                </a>
-                            </li>
-                            <li>
-                                <a onClick={() => openGenerateSQLTab()}>
-                                    <span className="icon">
-                                        <i className="fas fa-magic"></i>
-                                    </span>
-                                    &nbsp;Generate SQL
-                                </a>
-                            </li>
-                        </ul>
-                    </React.Fragment>
+                    <DatabaseSidebar />
                 }
                 {sidebarView === SidebarViewType.SETTINGS &&
-                    <React.Fragment>
-                        <p className="menu-label">
-                            Settings
-                        </p>
-                        <ul className={"menu-list " + styles.menuList}>
-                            {Constants.Build === 'server' &&
-                                <li>
-                                    <Link
-                                        to={Constants.APP_PATHS.SETTINGS_ACCOUNT.path}
-                                        className={location.pathname.startsWith(Constants.APP_PATHS.SETTINGS_ACCOUNT.path) ? 'is-active' : ''}>
-                                        Account
-                                    </Link>
-                                </li>
-                            }
-                            <li>
-                                <Link
-                                    to={Constants.APP_PATHS.SETTINGS_GENERAL.path}
-                                    className={location.pathname.startsWith(Constants.APP_PATHS.SETTINGS_GENERAL.path) ? 'is-active' : ''}>
-                                    General
-                                </Link>
-                            </li>
-                            <li>
-                                <Link
-                                    to={Constants.APP_PATHS.SETTINGS_ADVANCED.path}
-                                    className={location.pathname.startsWith(Constants.APP_PATHS.SETTINGS_ADVANCED.path) ? 'is-active' : ''}>
-                                    Advanced
-                                </Link>
-                            </li>
-                        </ul>
-                        {Constants.Build === 'server' && currentUser && currentUser.isRoot &&
-                            <React.Fragment>
-                                <p className="menu-label">
-                                    Manage Team
-                                </p>
-                                <ul className={"menu-list " + styles.menuList}>
-                                    <li>
-                                        <Link
-                                            to={Constants.APP_PATHS.SETTINGS_USERS.path}
-                                            className={location.pathname.startsWith(Constants.APP_PATHS.SETTINGS_USERS.path) ? 'is-active' : ''}>
-                                            Manage Users
-                                        </Link>
-                                    </li>
-                                    <li>
-                                        <Link
-                                            to={Constants.APP_PATHS.SETTINGS_ROLES.path}
-                                            className={location.pathname === Constants.APP_PATHS.SETTINGS_ROLES.path ? 'is-active' : ''}>
-                                            Manage Roles
-                                        </Link>
-                                    </li>
-                                </ul>
-                            </React.Fragment>
-                        }
-                        <p className="menu-label">
-                            Info
-                        </p>
-                        <ul className={"menu-list " + styles.menuList}>
-                            <li>
-                                <Link
-                                    to={Constants.APP_PATHS.SETTINGS_ABOUT.path}
-                                    className={location.pathname === Constants.APP_PATHS.SETTINGS_ABOUT.path ? 'is-active' : ''}>
-                                    About
-                                </Link>
-                            </li>
-                            <li>
-                                <Link
-                                    to={Constants.APP_PATHS.SETTINGS_SUPPORT.path}
-                                    className={location.pathname === Constants.APP_PATHS.SETTINGS_SUPPORT.path ? 'is-active' : ''}>
-                                    Support
-                                </Link>
-                            </li>
-                        </ul>
-                    </React.Fragment>
+                    <SettingSidebar />
                 }
             </div>
         </aside>
